Add an optional delay when powering off screen.
Add an option (power-off-delay, defaulted to 0) to delay stopping the compositor and powering off the screen to allow the shell to present a shutdown animation; without delay, the shell and-or greeter will get blocked  while trying to render such animation; this is due to the compositor stopping as USC receives the power state change before the shell.

This is mostly a workaround as ideally, the greeter and-or shell should coordinate when to shutdown the screen after they are given a chance to possibly show a shutdown animation.