Use get_options directly instead of usc::Server::the_options
It's not entirely obvious why this abstraction is here, but removing it
means that downstream code can use the same calling convention for both
`usc::Server` and `mir::Server` instances.  This should make it easier
to refactor out the various getter methods in `usc::Server` that depend
on access to options.