Check for --version flag in `main` method instead of SystemCompositor
This should make it easier to transfer from using `usc::Server` to using
`miral::MirRunner`, since it removes a key `SystemCompositor` dependency
on methods of the `Server` class.

When transitioning to `MirRunner` it should be possible to get the state
of the version flag from a callback used with a `CommandLineOption`.