Move usc::Server::the_dm_connection to usc::SystemCompositor internals
This is part of the work of unpicking the display-manager code from the
`usc::Server` class, so that the latter can be replaced with a simple
`mir::Server` instance.

A callback returning the `shared_ptr<SessionSwitcher>` instance is now
passed as a constructor parameter to `usc::SystemCompositor` (note that
it must be a callback and not the instance itself, because the instance
will not be initialized until after the `SystemCompositor` instance is
constructed).  This enables us to replace the `the_dm_connection` method
with direct instantiation of the `dm_connection` field from inside the
server init callback defined inside `usc::SystemCompositor::run`.  The
new code uses `the_session_switcher` directly rather than abstracting it
away behind a replacement `the_dm_message_handler`: the latter method of
`usc::Server` is simply removed.

Once `usc::Server` has been entirely replaced with a plain `mir::Server`
instance, it should be possible to further rework this code to replace
the `get_options()` calls with `CommandLineOption` instances.