Move ownership of SessionSwitcher outside usc::Server
This patch unpicks one of the tighter couplings of `usc::Server` and the
rest of the USC codebase, particularly with `usc::SystemCompositor`.

The `SessionSwitcher` instance forms the main co-dependency between the
`usc::Server` and `usc::SystemCompositor` class, since it is used in the
`the_dm_message_handler` method and thus the `the_dm_connection` method
used inside `usc::SystemCompositor`.  It is however first used in the
`usc::Server` constructor to override the window manager builder, and it
needs the server options to be available in order to construct it, since
it relies on both the `mir::Server` `--file` flag, and the `--spinner`
flag defined in USC.

In order to unpick this co-dependency, the `shared_ptr<SessionSwitcher>`
ownership has been moved to `main()`, where it is declared but not yet
initialized.  A reference to this is passed to the constructor of the
`usc::Server` class, which initializes it from within the window manager
override callback (!).  This is necessary in order to ensure the correct
sequence of events:

  * initializing the `SessionSwitcher` requires the `ExternalSpinner`
    instance to be initialized first

  * initializing the `ExternalSpinner` class requires access to the
    results of the `--file` command-line option, which can only be
    accessed via `mir::Server::get_options` (since it is a built-in
    flag defined internally by `mir::Server`)

  * `mir::Server::get_options` cannot be called until after the call
    to `mir::Server::apply_settings` at the end of the `usc::Server`
    constructor

  * the `apply_settings()` call will apply the window manager builder
    override, and therefore the callback, ensuring that the session
    switcher is initialized before the `usc::Server` constructor exits.

Since the `SessionSwitcher` stores the `shared_ptr<Spinner>` internally,
we are able to drop it as a field of both the `usc::SystemCompositor`
and `usc::Server` classes, since it is not needed to access directly.
This cleans up a bit of unnecessary `CachedPtr`-based lazy evaluation.
We also switch to using `get_options()` directly to access the results
of the `--file` and `--spinner` flags, in order to avoid depending on
custom methods of `usc::Server`.

Following this patch it should be possible to move the display-manager
methods (`the_dm_connection` and `the_dm_message_handler`) out of the
`usc::Server` class and into `usc::SystemCompositor` (which is where
they are actually used).