Make dbus_bus_address() an internal function of system_compositor.cpp
Now that all calling sites of this function have been moved over to the
`usc::SystemCompositor` class, we no longer need it to be accessible
from anywhere else.  Note that this might need to change in future if we
wish to break apart the `SystemCompositor` class itself, but for now it
suffices to keep it as an internal implementation detail.