Move creation of dbus event loop to usc::SystemCompositor
This patch removes the `usc::Server::the_dbus_event_loop` method and the
underlying `CachedPtr` field and instead creates the necessary instance
inside the init callback defined inside `usc::SystemCompositor::run`.

Since all consumers of the `shared_ptr<DBusEventLoop>` are created from
within the same callback, and assigned to fields of `SystemCompositor`,
this instance should stay valid for the lifetime of the latter class,
and hence for as long as needed by the application.