Move creation of Mir input configuration to usc::SystemCompositor
Since this is only used inside the init callback created inside the
`usc::SystemCompositor::run` method, and a copy of the shared pointer
will be kept inside the `UnityInputService` instance generated with it,
we do not need to provide this via a method of `usc::Server`, and we can
consequently dispense with the underlying `CachedPtr` field.

This patch leaves `usc::Server` with no custom fields or methods, only a
custom constructor.