Entirely replace usc::Server with mir::Server
Now that usc::Server has no custom methods or fields, it is possible to
entirely strip out that class and use `mir::Server` instead.  All of the
work done inside the old `usc::Server` constructor is now done instead
in the `main()` function.

This ought to leave us with a clearer sequence of events to unpick as we
transition towards using `miral::MirRunner`.

Note that all lambda closures specifying their capture rules as `[this]`
inside `usc::Server` have been updated to use `[&]` instead.  Obviously
we could use more specific captures, but this should make it simpler to
update code in future, and in particular to avoid hassle over capturing
by reference versus by value (which may have an impact for example when
we want to use values set by `CommandLineOption` callbacks).