Create display manager connection from main using command-line options
This patch reworks the creation of the `DMConnection` instance used in
USC, extracting it from the init callback added under the hood of the
`SystemCompositor` class, and placing it instead in `main`, where it is
instantiated using a collection of `CommandLineOption` instances.  Most
of the latter are implemented using `pre_init` just to ensure that the
resulting values are set correctly before the display manager connection
itself is instantiated; only the `--debug-without-dm` flag is not, since
its callback is the one inside which the instantiation actually happens.
(The use of `pre_init` for the others is probably overkill, but it can't
hurt to make sure that these values are handled as early as possible.)

This allows us to clean up some uses of `add_configuration_option` and
so reduce the number of direct dependencies on the `mir::Server` API.
We can also now remove the explicit `SystemCompositor` constructor, as
the class no longer needs to access the session switcher.