Implement --spinner flag using a pre-init CommandLineOption
Following earlier changes to how the socket file name is obtained, this
patch completes the process of removing any direct dependency on the
`mir::Server` API when constructing the `ExternalSpinner` instance.  A
`pre_init` command-line option is used to ensure that the `spinner_path`
variable will be set as early as possible in the app lifecycle.

The instantiation itself of course still occurs within a callback that
is passed directly to the server API, but this will be dealt with at a
later time.