Instantiate session_switcher in --spinner CommandLineOption callback
This patch ensures that the `spinner` and `session_switcher` can be set
using only the MirAL API.  Since `pre_init` is used, `session_switcher`
should be instantiated before the `override_the_window_manager_builder`
callback is invoked.  The latter callback is now limited to doing only
the `usc::WindowManager` instantiation that it needs to do.